+++ app/controllers/card_controller.rb
@@ -25,11 +25,12 @@ class CardController < ApplicationController
 
   def delete #PayjpとCardデータベースを削除します
     if @card.blank?
-        Payjp.api_key = ENV["PAYJP_PRIVATE_KEY"]
-        customer = Payjp::Customer.retrieve(@card.customer_id)
-        customer.delete
-        @card.delete
+      redirect_to action: "new"
     end
+      Payjp.api_key = ENV["PAYJP_PRIVATE_KEY"]
+      customer = Payjp::Customer.retrieve(@card.customer_id)
+      customer.delete
+      @card.delete
       redirect_to action: "new"
   end
