--- app/controllers/home_controller.rb
@@ -1,6 +1,0 @@
-class HomeController < ApplicationController
-  
-  def show
-  end
-
-end
