+++ app/uploaders/image_uploader.rb
@@ -7,8 +7,8 @@ class ImageUploader < CarrierWave::Uploader::Base
 
   
   #ローカルで画像テストはここ直して！！！！！！！！！！！！！！！！！！！！！！！！！！！
-  # storage :file
-  storage :fog  
+   storage :file
+  #storage :fog  
 
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
