+++ config/routes.rb
@@ -18,6 +18,13 @@ Rails.application.routes.draw do
   resources :home 
   resources :users
 
+  resources :buyers, only: :show do
+    collection do
+      post 'pay', to: 'buyers#pay'
+      get 'done', to: 'buyers#done'
+    end
+  end
+
   resources :card, only: [:new, :show] do
     collection do
       post 'pay', to: 'card#pay'
@@ -25,13 +32,6 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :purchase, only: [:index] do
-    collection do
-      post 'pay', to: 'purchase#pay'
-      get 'done', to: 'purchase#done'
-    end
-  end
 
-  resources :buyers
 end
